refactor: [1단계-문자열 계산기] 테스트용 코드 제거/불완전한 테스트 (nullAndEmpty에 관한 테스트가 불가능함)
