+++ src/main/java/calculator/domain/CalculationMethod.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.domain;
 
 import java.util.Arrays;
 import java.util.function.BiFunction;
@@ -8,7 +8,7 @@ public enum CalculationMethod {
     Sum("+", (num1, num2) -> Integer.parseInt(num1) + Integer.parseInt(num2)),
     Subtraction("-", (num1, num2) -> Integer.parseInt(num1) - Integer.parseInt(num2)),
     Multiplication("*", (num1, num2) -> Integer.parseInt(num1) * Integer.parseInt(num2)),
-    Division("/", (num1, num2) -> (int)Integer.parseInt(num1) / Integer.parseInt(num2));
+    Division("/", (num1, num2) -> (int) Integer.parseInt(num1) / Integer.parseInt(num2));
 
     private final String calculation;
     private final BiFunction<String, String, Integer> calculate;
@@ -18,23 +18,26 @@ public enum CalculationMethod {
         this.calculate = calculate;
     }
 
-    public int calculate(String num1, String num2) {
+    public int calculateApply(String num1, String num2) {
         try {
             calculate.apply(num1, num2);
-        }
-        catch (Exception e) {
+        } catch (Exception e) {
             throw new IllegalArgumentException("0으로 나눌 수 없습니다.");
         }
         return calculate.apply(num1, num2);
     }
 
+    public static int calculate(String calculation, String num1, String num2) {
+
+        CalculationMethod calculationMethod = find(calculation);
+        return calculationMethod.calculateApply(num1, num2);
+    }
+
     public static CalculationMethod find(String calculation) {
 
         return Arrays.stream(CalculationMethod.values())
-                    .filter(calculationMethod2 -> calculationMethod2.calculation.equals(calculation))
-                    .findFirst()
-                    .orElseThrow(() -> new IllegalArgumentException("+, -, *, / 중 하나여야 합니다!"));
+                .filter(calculationMethod2 -> calculationMethod2.calculation.equals(calculation))
+                .findFirst()
+                .orElseThrow(() -> new IllegalArgumentException("+, -, *, / 중 하나여야 합니다!"));
     }
-
-
 }
