+++ src/main/java/calculator/view/OutputView.java
@@ -1,4 +1,4 @@
-package outputview;
+package calculator.view;
 
 public class OutputView {
     public void printOutput(int output) {
