+++ src/test/java/calculatortest/CalculationMethodTest.java
@@ -1,6 +1,6 @@
 package calculatortest;
 
-import calculator.*;
+import calculator.domain.CalculationMethod;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -17,28 +17,28 @@ public class CalculationMethodTest {
     @Test
     @DisplayName("sum")
     public void sum() {
-        Assertions.assertThat(sum.calculate("1", "2")).isEqualTo(3);
+        Assertions.assertThat(sum.calculateApply("1", "2")).isEqualTo(3);
     }
 
     @Test
     @DisplayName("subtraction")
     public void subtraction() {
-        Assertions.assertThat(subtraction.calculate("1", "2")).isEqualTo(-1);
+        Assertions.assertThat(subtraction.calculateApply("1", "2")).isEqualTo(-1);
     }
 
     @Test
     @DisplayName("multiplication")
     public void multiplication() {
-        Assertions.assertThat(multiplication.calculate("1", "2")).isEqualTo(2);
+        Assertions.assertThat(multiplication.calculateApply("1", "2")).isEqualTo(2);
     }
 
     @Test
     @DisplayName("division")
     public void division() {
-        Assertions.assertThat(division.calculate("1", "2")).isEqualTo(0);
-        Assertions.assertThat(division.calculate("3", "2")).isEqualTo(1);
+        Assertions.assertThat(division.calculateApply("1", "2")).isEqualTo(0);
+        Assertions.assertThat(division.calculateApply("3", "2")).isEqualTo(1);
         Assertions.assertThatThrownBy(() -> {
-            division.calculate("3", "0");
+            division.calculateApply("3", "0");
         }).isInstanceOf(IllegalArgumentException.class);
     }
 }
