+++ src/test/java/calculatortest/CalculatorTest.java
@@ -1,12 +1,14 @@
 
 package calculatortest;
 
-import calculator.Calculator;
+import calculator.domain.Calculator;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import java.io.ByteArrayInputStream;
 
 public class CalculatorTest {
     //given
@@ -16,7 +18,9 @@ public class CalculatorTest {
     @DisplayName("calculator")
     @ValueSource(strings = {"4 + 3 * 2 / 1"})
     public void calculator(String string) {
-        Assertions.assertThat(calculator.calculate(string)).isEqualTo(14);
+        ByteArrayInputStream input = new ByteArrayInputStream(string.getBytes());
+        System.setIn(input);
+        Assertions.assertThat(calculator.calculate()).isEqualTo(14);
     }
 
     @ParameterizedTest
@@ -24,8 +28,10 @@ public class CalculatorTest {
     @NullAndEmptySource
     @ValueSource(strings = {"4@3", "4 @ 3", "4 / 0"})
     public void wrongCalculator(String string) {
+        ByteArrayInputStream input = new ByteArrayInputStream(string.getBytes());
+        System.setIn(input);
         Assertions.assertThatThrownBy(() -> {
-            calculator.calculate(string);
+            calculator.calculate();
         }).isInstanceOf(IllegalArgumentException.class);
     }
 }
